pyproject.toml
[project]
name = "api-hidro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "mvcastro", email = "mvcastro1975@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "types-requests>=2.32.4.20250913",
]

[project.scripts]
api-hidro = "api_hidro:main"

[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"
